Finally figured out the error in process sentinel
Turns out that if you do ";reload;compile" and there is an error in
the build script, once the error is fixed and you send "r", SBT will
recompile the project AND THEN EXIT.  So instead we'll send the
two commands separately.  It muddies the logic somewhat, but works
much better.